backend/api/requirements.txt
# AdData Dashboard API Requirements

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# ClickHouse driver
clickhouse-connect==0.7.10

# Configuration
pyyaml==6.0.1

# CORS
python-multipart==0.0.6

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
